ProtoWallet replaces WalletCrypto, ProtoWalletApi replaces WalletCryptoObject.
